Remarks3

+++ EStackException.h
@@ -10,13 +10,13 @@ public:
     EStackException(const char* arg_message)
     {
         message = new char[strlen(arg_message)+1];
-        for (int j = 0; j < strlen(arg_message) + 1; j++)
+        for (long long unsigned j = 0; j < strlen(arg_message) + 1; j++)
                     message[j] = arg_message[j];
     }
     EStackException(const EStackException& arg)
     {
         message = new char[strlen(arg.message)+1];
-        for (int j = 0; j < strlen(arg.message) + 1; j++)
+        for (long long unsigned j = 0; j < strlen(arg.message) + 1; j++)
                     message[j] = arg.message[j];
     }
     ~EStackException()
